feat: Docker Postgres added
Adds a docker folder with a docker database folder inside. This folder contains a .dockerfile that creates a postgres docker image that will be used  for development purposes. Also, to make it easier to call the commands, scripts have been added to package.json file, so it will be easier to build and run the docker image

sla